% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area.R
\name{get_area}
\alias{get_area}
\title{Get polygon(s) for an area of interest}
\usage{
get_area(name = "Australia", type = "eez", country_type = "country")
}
\arguments{
\item{name}{\code{character} name of the country or area. If \code{NULL} all areas are returned. If an incorrect \code{name} is input, a list of all possible names will be provided along with the error message.}

\item{type}{\code{character} the area type. Can be one of:
\itemize{
\item \code{eez}: Exclusive Economic Zone (EEZ; 200nm), differ slightly from the the \href{https://www.un.org/depts/los/convention_agreements/texts/unclos/part5.htm}{UN Convention on the Law of the Sea (UNCLOS)} definition because the archipelagic waters and the internal waters of a country are included
\item \verb{12nm}: 12 nautical miles zone (Territorial Seas), defined in \href{https://www.un.org/Depts/los/convention_agreements/texts/unclos/part2.htm}{UNCLOS}
\item \verb{24nm}: 14 nautical miles zone (Contiguous Zone), defined in \href{https://www.un.org/Depts/los/convention_agreements/texts/unclos/part2.htm}{UNCLOS}
\item \code{ocean}: Global Oceans and Seas as compiled by the Flanders Marine Data Centre. Names are: "Arctic Ocean", "Baltic Sea", "Indian Ocean", "Mediterranean Region", "North Atlantic Ocean", "North Pacific Ocean", "South Atlantic Ocean", "South China and Easter Archipelagic Seas", "South Pacific Ocean", and "Southern Ocean".
\item \code{countries}: country boundaries
}

More details on the marine areas can be found on the \href{https://marineregions.org/sources.php}{Marine Regions website}, and for land area, the \href{https://www.naturalearthdata.com/features/}{Natural Earth website}. Note that this function retrieves data from Natural Earth at the highest resolution (1:10m).}

\item{country_type}{\code{character} must be either \code{country} or \code{sovereign}. Some countries have many territories that it has jurisdiction over. For example, Australia, France and the U.K. have jurisdiction over many overseas islands. Using \code{sovereign} returns the main country and all the territories, whereas using \code{country} returns just the main country. More details about what is a country via the \code{rnaturalearth} package \href{https://cran.r-project.org/web/packages/rnaturalearth/vignettes/what-is-a-country.html}{vignette}}
}
\value{
'sf' object of the area requested
}
\description{
Marine and land areas can be obtained.  For marine areas, the \code{mrp_get} function from the \code{mregions2} package is used to retrieve an area (e.g. an EEZ) from \href{https://marineregions.org/gazetteer.php}{Marine Regions}. For land areas, the package \href{https://github.com/ropensci/rnaturalearth/}{\code{rnaturalearth}} is used.
}
\examples{
#Marine area examples:
if(require("mregions2")){
australia_mainland_eez <- get_area(name = "Australia")
plot(australia_mainland_eez["geometry"])

#this includes all islands that Australia has jurisdiction over:
australia_including_territories_eez <- get_area(name = "Australia", country_type = "sovereign")
plot(australia_including_territories_eez["geometry"])
}

#Land area examples:
if(require("rnaturalearth")){
australia_land <- get_area(name = "Australia", type = "countries")
plot(australia_land["geometry"])

#this includes all islands that Australia has jurisdiction over:
australia_land_and_territories <- get_area(name = "Australia", type = "countries", country_type = "sovereign")
plot(australia_land_and_territories["geometry"])
}
}
