% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialgridr-package.R
\docType{package}
\name{spatialgridr-package}
\alias{spatialgridr}
\alias{spatialgridr-package}
\title{spatialgridr: Grid spatial data for conservation planning purposes}
\description{
An R package that allows easy gridding of spatial data. Principally intended for spatial conservation planning uses.
}
\author{
\strong{Maintainer}: Jason Flower \email{jflower@ucsb.edu} (\href{https://orcid.org/0000-0002-6731-8182}{ORCID})

}
\keyword{internal}
