% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialgridr-package.R
\docType{package}
\name{spatialgridr-package}
\alias{spatialgridr}
\alias{spatialgridr-package}
\title{spatialgridr: Grid spatial data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R package that allows easy gridding of spatial data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://emlab-ucsb.github.io/spatialgridr/}
}

}
\author{
\strong{Maintainer}: Jason Flower \email{jflower@ucsb.edu} (\href{https://orcid.org/0000-0002-6731-8182}{ORCID})

}
\keyword{internal}
